+++ dt-common/src/monitor/mod.rs
@@ -3,9 +3,9 @@ use async_trait::async_trait;
 pub mod counter;
 pub mod counter_type;
 pub mod group_monitor;
-pub mod monitor_util;
 pub mod task_metrics;
 pub mod task_monitor;
+pub mod task_monitor_handle;
 
 #[allow(clippy::module_inception)]
 pub mod monitor;
