+++ dt-connector/src/checker/cdc_state.rs
@@ -512,7 +512,9 @@ mod tests {
     use crate::checker::check_log::CheckSummaryLog;
     use crate::rdb_router::RdbRouter;
     use async_trait::async_trait;
-    use dt_common::{monitor::task_monitor::TaskMonitorHandle, utils::limit_queue::LimitedQueue};
+    use dt_common::{
+        monitor::task_monitor_handle::TaskMonitorHandle, utils::limit_queue::LimitedQueue,
+    };
     use opendal::{services::Memory, Operator};
     use std::fs;
     use std::path::{Path, PathBuf};
@@ -575,7 +577,6 @@ mod tests {
                 monitor_task_id: "unit-test".to_string(),
                 base_sinker: crate::sinker::base_sinker::BaseSinker::new(
                     TaskMonitorHandle::default(),
-                    "unit-test".to_string(),
                     1,
                 ),
                 summary: CheckSummaryLog {
