+++ dt-task/src/parallelizer_util.rs
@@ -6,7 +6,7 @@ use super::task_util::TaskUtil;
 use dt_common::{
     config::{config_enums::ParallelType, sinker_config::SinkerConfig, task_config::TaskConfig},
     meta::redis::command::key_parser::KeyParser,
-    monitor::task_monitor::TaskMonitorHandle,
+    monitor::task_monitor_handle::TaskMonitorHandle,
     utils::redis_util::RedisUtil,
 };
 use dt_parallelizer::{
