+++ dt-task/src/sinker_util.rs
@@ -20,7 +20,7 @@ use dt_common::{
             redis_write_method::RedisWriteMethod,
         },
     },
-    monitor::task_monitor::TaskMonitorHandle,
+    monitor::task_monitor_handle::TaskMonitorHandle,
     rdb_filter::RdbFilter,
     utils::redis_util::RedisUtil,
 };
@@ -95,7 +95,7 @@ impl SinkerUtil {
         extractor_config: &ExtractorConfig,
         client: ConnClient,
         monitor: TaskMonitorHandle,
-        monitor_task_id: String,
+        _monitor_task_id: String,
         data_marker: Option<Arc<RwLock<DataMarker>>>,
         checker: Option<DataCheckerHandle>,
     ) -> anyhow::Result<Sinkers> {
@@ -138,11 +138,7 @@ impl SinkerUtil {
                         meta_manager: meta_manager.clone(),
                         router: router.clone(),
                         batch_size,
-                        base_sinker: BaseSinker::new(
-                            monitor.clone(),
-                            monitor_task_id.clone(),
-                            monitor_interval,
-                        ),
+                        base_sinker: BaseSinker::new(monitor.clone(), monitor_interval),
                         data_marker: data_marker.clone(),
                         replace,
                     };
@@ -174,11 +170,7 @@ impl SinkerUtil {
                         meta_manager: meta_manager.clone(),
                         router: router.clone(),
                         batch_size,
-                        base_sinker: BaseSinker::new(
-                            monitor.clone(),
-                            monitor_task_id.clone(),
-                            monitor_interval,
-                        ),
+                        base_sinker: BaseSinker::new(monitor.clone(), monitor_interval),
                         data_marker: data_marker.clone(),
                         replace,
                     };
@@ -199,11 +191,7 @@ impl SinkerUtil {
                         batch_size,
                         router: router.clone(),
                         mongo_client: mongo_client.clone(),
-                        base_sinker: BaseSinker::new(
-                            monitor.clone(),
-                            monitor_task_id.clone(),
-                            monitor_interval,
-                        ),
+                        base_sinker: BaseSinker::new(monitor.clone(), monitor_interval),
                     };
                     Self::push_checkable_sinker(&mut sub_sinkers, sinker, &checker);
                 }
@@ -247,11 +235,7 @@ impl SinkerUtil {
                         router: router.clone(),
                         producer,
                         avro_converter: avro_converter.clone(),
-                        base_sinker: BaseSinker::new(
-                            monitor.clone(),
-                            monitor_task_id.clone(),
-                            monitor_interval,
-                        ),
+                        base_sinker: BaseSinker::new(monitor.clone(), monitor_interval),
                     };
                     Self::push_sinker(&mut sub_sinkers, sinker);
                 }
@@ -274,11 +258,7 @@ impl SinkerUtil {
                     conflict_policy: conflict_policy.clone(),
                     filter: filter.clone(),
                     router,
-                    base_sinker: BaseSinker::new(
-                        monitor.clone(),
-                        monitor_task_id.clone(),
-                        monitor_interval,
-                    ),
+                    base_sinker: BaseSinker::new(monitor.clone(), monitor_interval),
                 };
                 Self::push_sinker(&mut sub_sinkers, sinker);
             }
@@ -300,11 +280,7 @@ impl SinkerUtil {
                     conflict_policy: conflict_policy.clone(),
                     filter: filter.clone(),
                     router,
-                    base_sinker: BaseSinker::new(
-                        monitor.clone(),
-                        monitor_task_id.clone(),
-                        monitor_interval,
-                    ),
+                    base_sinker: BaseSinker::new(monitor.clone(), monitor_interval),
                 };
                 Self::push_sinker(&mut sub_sinkers, sinker);
             }
@@ -343,11 +319,7 @@ impl SinkerUtil {
                             version,
                             method: method.clone(),
                             meta_manager: meta_manager.clone(),
-                            base_sinker: BaseSinker::new(
-                                monitor.clone(),
-                                monitor_task_id.clone(),
-                                monitor_interval,
-                            ),
+                            base_sinker: BaseSinker::new(monitor.clone(), monitor_interval),
                             data_marker: data_marker.clone(),
                             key_parser: KeyParser::new(),
                         };
@@ -364,11 +336,7 @@ impl SinkerUtil {
                             version,
                             method: method.clone(),
                             meta_manager: meta_manager.clone(),
-                            base_sinker: BaseSinker::new(
-                                monitor.clone(),
-                                monitor_task_id.clone(),
-                                monitor_interval,
-                            ),
+                            base_sinker: BaseSinker::new(monitor.clone(), monitor_interval),
                             data_marker: data_marker.clone(),
                             key_parser: KeyParser::new(),
                         };
@@ -389,11 +357,7 @@ impl SinkerUtil {
                         statistic_type: statistic_type.clone(),
                         data_size_threshold,
                         freq_threshold,
-                        base_sinker: BaseSinker::new(
-                            monitor.clone(),
-                            monitor_task_id.clone(),
-                            monitor_interval,
-                        ),
+                        base_sinker: BaseSinker::new(monitor.clone(), monitor_interval),
                     };
                     Self::push_sinker(&mut sub_sinkers, sinker);
                 }
@@ -447,11 +411,7 @@ impl SinkerUtil {
                         password,
                         batch_size,
                         meta_manager,
-                        base_sinker: BaseSinker::new(
-                            monitor.clone(),
-                            monitor_task_id.clone(),
-                            monitor_interval,
-                        ),
+                        base_sinker: BaseSinker::new(monitor.clone(), monitor_interval),
                         sync_timestamp: Utc::now().timestamp_millis(),
                         hard_delete: false,
                     };
@@ -518,11 +478,7 @@ impl SinkerUtil {
                         username,
                         password,
                         batch_size,
-                        base_sinker: BaseSinker::new(
-                            monitor.clone(),
-                            monitor_task_id.clone(),
-                            monitor_interval,
-                        ),
+                        base_sinker: BaseSinker::new(monitor.clone(), monitor_interval),
                         sync_timestamp: Utc::now().timestamp_millis(),
                     };
                     Self::push_sinker(&mut sub_sinkers, sinker);
@@ -569,11 +525,7 @@ impl SinkerUtil {
                         meta_manager,
                         router: router.clone(),
                         reverse,
-                        base_sinker: BaseSinker::new(
-                            monitor.clone(),
-                            monitor_task_id.clone(),
-                            monitor_interval,
-                        ),
+                        base_sinker: BaseSinker::new(monitor.clone(), monitor_interval),
                     };
                     Self::push_sinker(&mut sub_sinkers, sinker);
                 }
@@ -618,11 +570,7 @@ impl SinkerUtil {
                         batch_memory_bytes: batch_memory_mb * 1024 * 1024,
                         s3_config: s3_config.clone(),
                         s3_client: s3_client.clone(),
-                        base_sinker: BaseSinker::new(
-                            monitor.clone(),
-                            monitor_task_id.clone(),
-                            monitor_interval,
-                        ),
+                        base_sinker: BaseSinker::new(monitor.clone(), monitor_interval),
                         schema,
                         tb,
                         reverse_router: reverse_router.clone(),
@@ -633,11 +581,7 @@ impl SinkerUtil {
                         batch_size,
                         s3_config: s3_config.clone(),
                         s3_client: s3_client.clone(),
-                        base_sinker: BaseSinker::new(
-                            monitor.clone(),
-                            monitor_task_id.clone(),
-                            monitor_interval,
-                        ),
+                        base_sinker: BaseSinker::new(monitor.clone(), monitor_interval),
                         conn_pool: conn_pool.clone(),
                         extract_type: config.extractor_basic.extract_type.clone(),
                     };
@@ -646,11 +590,7 @@ impl SinkerUtil {
                         url: url.to_string(),
                         meta_manager,
                         batch_size,
-                        base_sinker: BaseSinker::new(
-                            monitor.clone(),
-                            monitor_task_id.clone(),
-                            monitor_interval,
-                        ),
+                        base_sinker: BaseSinker::new(monitor.clone(), monitor_interval),
                         conn_pool: conn_pool.clone(),
                         router: router.clone(),
                         pusher,
@@ -698,11 +638,7 @@ impl SinkerUtil {
                         batch_memory_bytes: batch_memory_mb * 1024 * 1024,
                         s3_config: s3_config.clone(),
                         s3_client: s3_client.clone(),
-                        base_sinker: BaseSinker::new(
-                            monitor.clone(),
-                            monitor_task_id.clone(),
-                            monitor_interval,
-                        ),
+                        base_sinker: BaseSinker::new(monitor.clone(), monitor_interval),
                         schema,
                         tb,
                         reverse_router: reverse_router.clone(),
@@ -733,11 +669,7 @@ impl SinkerUtil {
                         batch_size,
                         s3_config: s3_config.clone(),
                         s3_client: s3_client.clone(),
-                        base_sinker: BaseSinker::new(
-                            monitor.clone(),
-                            monitor_task_id.clone(),
-                            monitor_interval,
-                        ),
+                        base_sinker: BaseSinker::new(monitor.clone(), monitor_interval),
                         conn_pool: conn_pool.clone(),
                         extract_type: config.extractor_basic.extract_type.clone(),
                     };
@@ -767,11 +699,7 @@ impl SinkerUtil {
                     filter,
                     router,
                     engine,
-                    base_sinker: BaseSinker::new(
-                        monitor.clone(),
-                        monitor_task_id.clone(),
-                        monitor_interval,
-                    ),
+                    base_sinker: BaseSinker::new(monitor.clone(), monitor_interval),
                 };
                 Self::push_sinker(&mut sub_sinkers, sinker);
             }
