+++ dt-tests/tests/pg_to_pg/snapshot/parallel_test/resume_db_test/task_config.ini
@@ -6,6 +6,7 @@ db_type=pg
 extract_type=snapshot
 url={pg_extractor_url}
 parallel_size=3
+parallel_type=chunk
 batch_size=3
 
 [sinker]
