+++ dt-tests/tests/pg_to_pg/snapshot/parallel_test/resume_log_test/task_config.ini
@@ -3,6 +3,7 @@ db_type=pg
 extract_type=snapshot
 url={pg_extractor_url}
 parallel_size=3
+parallel_type=chunk
 batch_size=3
 
 [sinker]
